requirements.txt
annotated-types==0.7.0
anyio==4.4.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
beanie==1.26.0
certifi==2024.7.4
cffi==1.16.0
click==8.1.7
colorama==0.4.6
dnspython==2.6.1
email-validator==2.2.0
fastapi==0.111.1
fastapi-cli==0.0.4
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
jinja2==3.1.4
lazy-model==0.2.0
markdown-it-py==3.0.0
markupsafe==2.1.5
mdurl==0.1.2
motor==3.5.1
pycparser==2.22
pydantic==2.8.2
pydantic-core==2.20.1
pydantic-settings==2.4.0
pygments==2.18.0
pyjwt==2.8.0
pymongo==4.8.0
python-decouple==3.8
python-dotenv==1.0.1
python-multipart==0.0.9
pyyaml==6.0.1
rich==13.7.1
shellingham==1.5.4
sniffio==1.3.1
starlette==0.37.2
toml==0.10.2
typer==0.12.3
typing-extensions==4.12.2
uvicorn==0.30.3
watchfiles==0.22.0
websockets==12.0
